Drop styled components for now
This is interfering with filetype plugin config: the filetype for
typescript files remains "typescript", but the keybinding for filetype
config is javascript. Chances are it's loading both filetype config
files, and the js one comes first with "nnoremap" for the config file
bindings, which prevents the ts file from adding its bindings.