Fix direnv load with multiline prompt after keybinds
Echo creates a blank line. Without this, the first line of the prompt
overwrites the last line of the direnv load.

Repainting forces the prompt to display on this line. Without this, the
prompt doesn't display if direnv loads after the keybinding. This only
affected the fishrc and nvimrc functions.